+++ app/models/thredded/user_preference.rb
@@ -8,7 +8,7 @@ module Thredded
              foreign_key: :user_id,
              inverse_of: :user_preference
     validates :user_id, presence: true
-    serialize :notifications_for_followed_topics, TruthyHashSerializer
-    serialize :notifications_for_private_topics, TruthyHashSerializer
+    serialize :notifications_for_followed_topics, TruthyHash::Serializer
+    serialize :notifications_for_private_topics, TruthyHash::Serializer
   end
 end
