+++ spec/commands/thredded/notify_private_topic_users_spec.rb
@@ -25,7 +25,7 @@ module Thredded
           create(
             :user_preference,
             user: @joel,
-            notifications_for_private_topics: TruthyHashSerializer.create('email' => false)
+            notifications_for_private_topics: TruthyHash.create('email' => false)
           )
           recipients = NotifyPrivateTopicUsers.new(post).targeted_users(notifier)
           expect(recipients).not_to include(@joel)
